Add Collision detection
Use collision detection to prevent solid bodies to overlap. Solution is
not perfect, object are not always able to touch each other. The
collision prevent the next move if it collide but there is not go as
far as you can. The display also use rounded value for positions that
emplify the gap effect.