Add Spwaner to create some mobs
Create some spot for entities to pop. The spwaner take a Factory and
some options to adjust spot capacity, regen time etc.

The Spwaner have to register to the GameEntityStore in order to be
updated and notified when entities needs to be cleaned up.

There are some bugs that need to be solved. Spwaners may stop poping
mobs or generate more than capacity.